DjangoProject/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
Jinja2==3.1.3
Markdown==3.6
requests==2.31.0
urllib3>=2.2.2
pandas>=1.5.0
openpyxl>=3.0.0
chardet>=5.2.0
importlib-metadata>=4.6.0
Pillow>=10.0.0
